chore(refactor): add ingress event type and refactor timeline
