+++ backend/controller/sql/schema/20240917015216_add_ingress_event_type.sql
@@ -1,0 +1,6 @@
+-- migrate:up
+
+ALTER TYPE event_type ADD VALUE IF NOT EXISTS 'ingress';
+
+-- migrate:down
+
