+++ deployment/base/db-migrate/kustomization.yml
@@ -26,4 +26,5 @@ configMapGenerator:
   - ./schema/20240913041619_encrypted_topic_events_payload.sql
   - ./schema/20240916015906_remove_runner_state.sql
   - ./schema/20240916190209_rename_controller_to_controllers.sql
+  - ./schema/20240917015216_add_ingress_event_type.sql
   - ./schema/20240917062716_change_deployments_index.sql
